fix: check for existence of backed up keybinding config before restoring

+++ packages/plugin-core/src/test/suite-integ/Extension.test.ts
@@ -99,8 +99,10 @@ suite("Extension", function () {
     afterHook: async () => {
       homeDirStub.restore();
       fs.removeSync(keybindingConfigPath);
-      fs.copyFileSync(backupKeybindingConfigPath, keybindingConfigPath);
-      fs.removeSync(backupKeybindingConfigPath);
+      if (fs.existsSync(backupKeybindingConfigPath)) {
+        fs.copyFileSync(backupKeybindingConfigPath, keybindingConfigPath);
+        fs.removeSync(backupKeybindingConfigPath);
+      }
     },
     noSetInstallStatus: true,
   });
